code review: review OffsetIndex
